Extract command parsing logic to separate class
Extract command parsing to separate class

Command processing is all performed in the Duke class.
This results in a very long file, with no clear separation of
responsibilities. It also makes the file messy with UI and parsing
logic.

Let's
- Extract command parsing to the Parser class
- Hide command processing logic from the Duke class
